+++ src/storage/inMemory/queue.storage.js
@@ -1,4 +1,4 @@
-import { Storage } from '../storage/Storage.js';
+import { Storage } from './Storage.js';
 
 export class Queue extends Storage {
     async getCurrentKey() {
